// src/modules/veiculos/enums/EEspecieVeiculo.ts
enum EEspecieVeiculo {
  Motocicleta = 1,
  Triciclo = 2,
  Quadriciclo = 3,
  Automóvel = 4,
  Caminhonete = 5,
  Caminhão = 6,
  Camioneta = 7,
  Microônibus = 8,
  Ônibus = 9,
  'Reboque ou Semi-reboque' = 10,
  Utilitário = 11,
  Trator = 12,
  Outros = 13,
}

export default EEspecieVeiculo;